+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## unreleased
 
+### Fixed
+
+ - Initialise `.data.dummy` to `\0` for empty variants. Previously, it was uninitialised.
+
 ## 1.6.1 - 2021-12-09
 
 ### Fixed